revision day, naprawa projektu, usuniecie cachowania(bo zbyt komplikowalo projekt), poczatki ze spritami dla body.
Dodanie podstawowego rzucania podsatwowego ataku.

+++ core/src/com/brotherhood/wizards/spells/dto/SpellDTO.java
@@ -81,8 +81,8 @@ public class SpellDTO {
         this.defencePoints = defencePoints;
     }
 
-    public double getSpeed() {
-        return speed;
+    public float getSpeed() {
+        return (float) speed;
     }
 
     public void setSpeed(double speed) {
